Fix warning about implicit cast

+++ Source/UI/Text.cpp
@@ -66,7 +66,7 @@ Text::Text() {
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            (GLuint)face->glyph->advance.x
         };
         characters.insert(std::pair<GLchar, Character>(i, character));
     }
